Extract a Price component
